requirements.txt
streamlit==1.39.0
moviepy==1.0.3
SQLAlchemy==2.0.35
bcrypt==4.2.0
youtube-transcript-api==0.6.2
SpeechRecognition==3.10.4
pydub==0.25.1
pytube==15.0.0
streamlit-cookies-manager==0.2.0
pymysql==1.1.1
indic-transliteration==2.3.68
google-generativeai==0.8.3
pypdf==5.0.1
python-docx==1.1.2
selenium==4.31.1